refactor(adhd): перенос функционала AdhdHub в TelegramusHub
