+++ Program.cs
@@ -6,6 +6,7 @@ using MARS.Server.CustomLoggers.TelegramLogger;
 using MARS.Server.DataBaseContext;
 using MARS.Server.Exstensions;
 using MARS.Server.Hubs;
+using MARS.Server.Hubs.AudioControllerHub;
 using MARS.Server.Services.CinemaQueue;
 using MARS.Server.Services.CommandExecutor;
 using MARS.Server.Services.Configuration;
@@ -296,9 +297,8 @@ public static class Program
         app.MapHub<TunaHub>("/hubs/tuna");
         //app.MapHub<SoundBarHub>("/hubs/soundbar");
         app.MapHub<ScoreboardHub>("/hubs/scoreboard");
-        app.MapHub<AdhdHub>("/hubs/adhd");
         app.MapHub<VoiceRecognitionHub>("/hubs/tts");
-        app.MapHub<Hubs.AudioControllerHub.AudioControllerHub>("/hubs/audio-controller");
+        app.MapHub<AudioControllerHub>("/hubs/audio-controller");
         app.AddLogerHub();
         if (IsUseSoundRequest)
         {
